+++ src/app/logout/page.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 
 const page = () => {
   return (
-    <div>Sing Out</div>
+    <div>Logout</div>
   )
 }
